aaaaaaaaaa

+++ src/Utils/messages-media.ts
@@ -415,13 +415,15 @@ export const prepareStream = async(
 	logger?.debug('fetched media stream')
 
 	let bodyPath: string | undefined
+	let didSaveToTmpPath = false
 	try {
 		const buffer = await toBuffer(stream)
 		if(type === 'file') {
 			bodyPath = (media as any).url
 		} else if(saveOriginalFileIfRequired) {
-			bodyPath = join(getTmpFilesDirectory(), mediaType + generateMessageIDV2())
+			bodyPath = join(getTmpFilesDirectory(), mediaType + generateMessageID())
 			writeFileSync(bodyPath, buffer)
+			didSaveToTmpPath = true
 		}
 
 		const fileLength = buffer.length
@@ -436,17 +438,20 @@ export const prepareStream = async(
 			fileLength,
 			fileSha256,
 			fileEncSha256: undefined,
-			bodyPath
+			bodyPath,
+			didSaveToTmpPath
 		}
 	} catch (error) {
 		// destroy all streams with error
 		stream.destroy()
 
-		try {
+		if(didSaveToTmpPath) {
+			try {
 				await fs.unlink(bodyPath!)
 			} catch(err) {
 				logger?.error({ err }, 'failed to save to tmp path')
 			}
+		}
 
 		throw error
 	}
@@ -463,35 +468,24 @@ export const encryptedStream = async(
 
 	const mediaKey = Crypto.randomBytes(32)
 	const { cipherKey, iv, macKey } = await getMediaKeys(mediaKey, mediaType)
+	const encWriteStream = new Readable({ read: () => {} })
 
-	const encFilePath = join(
-		getTmpFilesDirectory(),
-		mediaType + generateMessageIDV2() + '-enc'
-	)
-	const encFileWriteStream = createWriteStream(encFilePath)
-
-	let originalFileStream: WriteStream | undefined
-	let originalFilePath: string | undefined
-
-	if(saveOriginalFileIfRequired) {
-		originalFilePath = join(
-			getTmpFilesDirectory(),
-			mediaType + generateMessageIDV2() + '-original'
-		)
-		originalFileStream = createWriteStream(originalFilePath)
+	let bodyPath: string | undefined
+	let writeStream: WriteStream | undefined
+	let didSaveToTmpPath = false
+	if(type === 'file') {
+		bodyPath = (media as any).url
+	} else if(saveOriginalFileIfRequired) {
+		bodyPath = join(getTmpFilesDirectory(), mediaType + generateMessageID())
+		writeStream = createWriteStream(bodyPath)
+		didSaveToTmpPath = true
 	}
 
 	let fileLength = 0
 	const aes = Crypto.createCipheriv('aes-256-cbc', cipherKey, iv)
-	const hmac = Crypto.createHmac('sha256', macKey!).update(iv)
-	const sha256Plain = Crypto.createHash('sha256')
-	const sha256Enc = Crypto.createHash('sha256')
-
-	const onChunk = (buff: Buffer) => {
-		sha256Enc.update(buff)
-		hmac.update(buff)
-		encFileWriteStream.write(buff)
-	}
+	let hmac = Crypto.createHmac('sha256', macKey!).update(iv)
+	let sha256Plain = Crypto.createHash('sha256')
+	let sha256Enc = Crypto.createHash('sha256')
 
 	try {
 		for await (const data of stream) {
@@ -510,62 +504,67 @@ export const encryptedStream = async(
 				)
 			}
 
-			if(originalFileStream) {
-				if(!originalFileStream.write(data)) {
-					await once(originalFileStream, 'drain')
+			sha256Plain = sha256Plain.update(data)
+			if(writeStream) {
+				if(!writeStream.write(data)) {
+					await once(writeStream, 'drain')
 				}
 			}
 
-			sha256Plain.update(data)
 			onChunk(aes.update(data))
 		}
 
 		onChunk(aes.final())
 
 		const mac = hmac.digest().slice(0, 10)
-		sha256Enc.update(mac)
+		sha256Enc = sha256Enc.update(mac)
 
 		const fileSha256 = sha256Plain.digest()
 		const fileEncSha256 = sha256Enc.digest()
 
-		encFileWriteStream.write(mac)
+		encWriteStream.push(mac)
+		encWriteStream.push(null)
 
-		encFileWriteStream.end()
-		originalFileStream?.end?.()
+		writeStream?.end()
 		stream.destroy()
 
 		logger?.debug('encrypted data successfully')
 
 		return {
 			mediaKey,
-			originalFilePath,
-			encFilePath,
+			encWriteStream,
+			bodyPath,
 			mac,
 			fileEncSha256,
 			fileSha256,
-			fileLength
+			fileLength,
+			didSaveToTmpPath
 		}
 	} catch(error) {
 		// destroy all streams with error
-		encFileWriteStream.destroy()
-		originalFileStream?.destroy?.()
+		encWriteStream.destroy()
+		writeStream?.destroy()
 		aes.destroy()
 		hmac.destroy()
 		sha256Plain.destroy()
 		sha256Enc.destroy()
 		stream.destroy()
 
-
-		try {
-			await fs.unlink(encFilePath)
-			if(originalFilePath) {
-				await fs.unlink(originalFilePath)
+		if(didSaveToTmpPath) {
+			try {
+				await fs.unlink(bodyPath!)
+			} catch(err) {
+				logger?.error({ err }, 'failed to save to tmp path')
 			}
-		} catch(err) {
-			logger?.error({ err }, 'failed deleting tmp files')
 		}
 
 		throw error
+	}
+
+	function onChunk(buff: Buffer) {
+		sha256Enc = sha256Enc.update(buff)
+		hmac = hmac.update(buff)
+		encWriteStream.push(buff)
 	}
 }
